+++ packages/components/src/components/telekom/app-header/app-header.tsx
@@ -381,6 +381,7 @@ export class Header {
                     href={this.logoHref}
                     logoTitle={this.logoTitle}
                     onClick={this.logoClick}
+                    focusable={this.scrolled}
                   ></app-logo>
                 </div>
                 <div class="header__nav-menu-wrapper">