+++ packages/storybook-vue/stories/3_components/notification-message/ScaleNotificationMessage.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <slot name="text" />
+  </div>
+</template>
+
+<script>
+export default {
+  props: {
+    variant: {
+      type: 'error' | 'warning' | 'success' | 'informational',
+      default: 'informational',
+    },
+    dismissible: { type: Boolean, default: false },
+    opened: { type: Boolean },
+    timeout: { type: Boolean | Number, default: false },
+  },
+};
+</script>