+++ packages/storybook-vue/stories/3_components/rating-stars/ScaleRatingStars.vue
@@ -10,12 +10,15 @@
     :aria-label-translation="ariaLabelTranslation"
     :label="label"
     :readonly="readonly"
+    :info-text="infoText"
+    :hide-label="hideLabel"
     @scaleChange="scaleChange"
   >
   </scale-rating-stars>
 </template>
 
 <script>
+import { action } from "@storybook/addon-actions";
 export default {
   props: {
     maxRating: { type: Number, default: 5 },
@@ -26,7 +29,9 @@ export default {
     size: { type: String, default: 'large' },
     disabled: { type: Boolean, default: false },
     ariaLabelTranslation: { type: String },
-    label: { type: String },
+    label: { type: String, default: 'Rating' },
+    infoText: { type: String },
+    hideLabel: { type: Boolean, default: false },
     readonly: { type: Boolean, default: false },
   },
   methods: {